Update Quke framework to latest version
We hit an issue with the tests running due to Chrome updating to version 74.

The instance of Chromedriver helper installed with the current version of Quke only supported versions 70 to 73. So to keep things in sync we needed to update Quke to a new version in order to pull down a newer version of Chromedriver helper.

So this change updates Quke and all its dependencies to the latest version.