Update calls to wait_for in SitePrism
The older version of SitePrism being used in the previous version of Quke supported a `wait_for` call on page object elements.

This appears to have changed in the latest versions of SitePrism to be `wait_until_element_visible`.

So this change updates any reference to `wait_for_element` to use the new syntax.