Migrate Login tests
The existing tests have one that confirms login into the external app works. We migrate that along with adding one for the internal app. It made sense to us to have the same test for both apps. Plus it allows us to confirm all is working for both apps with our first test.