Migrate external bulk upload test
Main thing to highlight about this test is it does not pass. So, we have migrated it as `skipped()` to avoid a false positive in the reports.

This has been failing for as long as we have been using the tests at this point

```javascript
cy.get('p > a').click()
```

This element is nowhere to be seen in the page. Our problem is the rest of the code does not relate to anything we see in the external UI. So, at this point we've no idea what this test was supposed to be doing.

But we migrate it anyway as it's on our to-do list of things to look into and resolve.