Migrate notify callback test
This test has to replicate the GOV.UK notify service hitting an endpoint we've provided for callbacks. For whatever reason, the previous team brought in [axios](https://www.npmjs.com/package/axios) to do this, even though sending HTTP requests is the core of Cypress!

So, most of the work was done in the initial commit (can you tell some of these tests were migrated awhile ago!?) but we've managed to strip out all the additional dependencies the old code was reliant on. Even said goodbye to [Lodash](https://www.npmjs.com/package/lodash)!