Add a body to our POST requests
Testing against the environments was failing because the tear-down requests were returning an error. After investigating the issue we have found the requests are being rejected by the [Silverline WAF](https://www.f5.com/cloud/products/distributed-cloud-waf) used in our AWS environments because they don't have a body. For 'reasons' there is a policy to automatically reject POST requests with no body.

We're getting that changed but in the meantime adding a dummy body bypasses the issue.