Add a body to our POST requests (#6)
Testing against the environments was failing because the tear-down requests were returning an error. After investigating the issue we have found the requests are being rejected by the [Silverline WAF](https://www.f5.com/cloud/products/distributed-cloud-waf) used in our AWS environments because they don't have a body. For 'reasons' there is a policy to automatically reject POST requests with no body.

We're getting that changed but in the meantime adding a dummy body bypasses the issue. In getting this working we also found that the test data tear-down and setup actions take longer than the default 30 seconds in some of our non-prod environments. So, this change also includes an increase in the response timeout for those specific requests.