Migrate sharing licence access external tests
This migrates 2 external tests. Both are to do with an existing user with access to a licence granting access to another.

In the first test the other user already exists in the system. In the second the user is new, which involves an email being sent to them and the test therefore needing to follow the link to register.

This is the second time we've had to manipulate the body of the last notification received. We're jumping the gun on the [rule of 3](https://en.wikipedia.org/wiki/Rule_of_three_(computer_programming)) but as it was already a gnarly bit of code we using this as an excuse to clean it up.