Migrate user registration external test
This migrates the external user registration test. When we reviewed what was there we found most of it was duplicating validation that had already been checked else where. It could also be argued what this is testing is checked in `cypress/e2e/external/sharing-access-with-new-user.cy.js` but we've retained it to be on the safe side.

So, in this change we create a `login/` folder and move the existing `cypress/e2e/external/login-logout.cy.js` in there along with a new validation of the login page test.

We then separate out the journey from the rest of the validation testing. The end result though is `repos/water-abstraction-ui/cypress/integration/external/user-registration.spec.js` is migrated.