Migrate user registration external test (#18)
This migrates the external user registration test. When we reviewed what was there we found most of it duplicated validation that had already been checked elsewhere. It could also be argued that this testing is checked in `cypress/e2e/external/sharing-access-with-new-user.cy.js` but we've retained it to be on the safe side.

What we have done is follow the pattern we used for **reset-password** where we split the validation tests from the journey test, to make what we're checking clearer.

Whilst we were doing this, we realised we should really be doing the same for the **login**. 

So, in this change, we also create a `login/` folder and move the existing `cypress/e2e/external/login-logout.cy.js` in there along with a new validation of the login page test.