Migrate view licences test
This is the final external UI acceptance test to migrate.

TBH, it is checking a page most of the other journeys go through. But we're migrating at this time to ensure there are no concerns with something being dropped by the the wider team. We can review properly once everything is migrated.