Migrate view licences test (#22)
This is the final external UI acceptance test to migrate.

TBH, it is checking a page most of the other journeys go through. But we're migrating now to ensure there are no concerns with something being dropped by the wider team. We can review everything once it is all migrated.