Move create user test
The tests we are migrating were all in a folder named `user/`. But as we're migrating them we are seeing that doesn't really make much sense.

So, left as is `create-user.cy.js` would be the only test in `user/`. To us, it makes more sense to move it to the root.