Migrate internal user tests (#28)
This change migrates the legacy internal user tests to this project. The legacy tests were

- create-user.spec.js
- reset-password.spec.js
- search-for-licence.spec.js
- view-contacts.spec.js

The tests we are migrating were all in a folder named `user/`. But as we're migrating them we are seeing that doesn't really make much sense. Most we're just placing in the root as they are standalone tests.

In the legacy reset-password test validation and the journey were mixed in. To make a clearer distinction about what we're testing we have split them into 2; journey and validation.

This is the same thing we did for the internal reset password migration.