Migrate abstraction alert tests
This change migrates the legacy water abstraction alert acceptance tests. These can be found in [cypress/integration/internal/waa](https://github.com/DEFRA/water-abstraction-ui/tree/main/cypress/integration/internal/waa).

Again as previously found, the existing tests contain a mix of testing the journey and _some_ validations. To make it easier to asses their value going forward the 2 have been split.

In this particular case the legacy code appeared to contain 2 distinct tests; `gauging-stations-and-water-abstraction-alerts-mbod.spec` and `gauging-stations-and-water-abstraction-alerts.spec`. However, after review they are almost identical. The only difference is in the choice of unit of measure for the threshold, **mBOD** or **Ml/d**. Whichever you select makes no difference to the subsequent journey. So, as part of the migration we have dropped the test for **Ml/d** as it is adding no value.