Migrate paper return form tests (#34)
This is a partial migration of the legacy acceptance test `paper-forms-flow.ci.spec`.

Though the test is ostensibly about sending paper return forms for specified licences within is an exhaustive check of the address lookup functionality.

To make things easier to review and maintain in the future we're splitting the test up. This change covers the paper return form process. Again, we've split the parts looking at validation from the journey.

We'll migrate the address lookup elements in a subsequent change.