Migrate address entry tests (#36)
This is the remainder of the migration of the legacy acceptance test `paper-forms-flow.ci.spec`.

Though the test is ostensibly about sending paper return forms for specified licences within it is an exhaustive check of the address entry functionality.

To make things easier to review and maintain in the future we split the test up. [Migrate paper return form tests](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/34) covered the paper return forms. This change migrates all the address entry tests.