Migrate charge information tests
This covers the migration of what we found in `internal/chargeinformation` in the legacy tests, having dealt with [user permission](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/40) and [licence agreement](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/41) tests.

We migrated `remove-charge-info-workflow.spec.js` into its own `workflow/` folder as it was actually testing the workflow functionality rather than charge information.

The remaining tests were probably the worst we've had to deal with so far. They were

- sroc-charge-version-adjustments-option-no.spec.js
- sroc-charge-version-adjustments-option-yes.spec.js
- sroc-charge-version.spec.js
- sroc-edit-submit-review-view-charge-version.spec.js
- sroc-select-multiple-charge-version.spec.js

5 tests that all repeat the exact same validation checks and have very little to differentiate them. They are all go through the same set up new charge version journey with very, _very_ minor differences. `sroc-charge-version.spec.js`, for example, create a charge version that adds the same adjustment used in `sroc-charge-version-adjustments-option-yes.spec.js` and 2 others. `sroc-edit-submit-review-view-charge-version.spec.js` is a duplicate of most of them only it goes a few steps further in the journey (and nothing gets edited!)

So, we have condensed them down into 2 tests;

- sroc/journey.cy.js
- sroc/validation.cy.js

`journey.cy.js` includes adding an adjustment and an additional element. It also includes reviewing and approving the charge version. It omits the validation and asserting values appear where they should be as that is covered by `validation.cy.js`. There was no validation checking of the screens used to add a charge element so that part is omitted from the journey. The same goes for the review and approve screens.

We cover everything the legacy tests were doing including additional checking of values shown in the pages. But we remove the duplication which reduces the overall run time and will make it easier to maintain in future.