Migrate change user permissions test (#40)
The legacy test `change-user-permissions.spec.js` was found in the `internal/chargeinformation` folder when we were about to migrate the other tests in there.

Clearly, it has nothing to do with charge information so we decided to migrate it separately. In doing so it gave us at least 2 user tests which meant a `user/` folder was needed. We then had another think about `login-logout.cy.js` and decided to give it a home in the `user/` folder.