Migrate licence agreement tests
When looking to migrate the tests in `internal/chargeinformation/` we found `create-view-edit-delete-licence-agreements.spec.js`.

Managing licence agreements is done from the charge information tab in a licence. But we felt it is a feature in its own right and as such should be in a separate folder.

On top of that there are 3 functions of a licence agreement being tested in one file. We understand why; the test data does not include a licence agreement so one has to be created through the UI. But these are relatively quick tests to run so when migrating we choose to split it into 3 distinct tests

- delete-licence-agreement.cy.js
- end-licence-agreement.cy.js
- new-licence-agreement.cy.js

We just need to get through all the migrations and then we can start being a bit cleverer with the set up data!