Migrate legacy Licence transfer test
We're onto the last folder of legacy tests; `internal/billing`. No surprise there are tests that have little to do with billing and would be better located elsewhere. The first of these was [mark-licence-for-supplementary-billing.spec.js](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/43).  The second is `license-transfer-to-new-billing-account.spec.js`.

This is _another_ duplicate test of adding a charge information (see [Migrate charge information tests](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/42) for our thoughts on those tests!)

The only change in this one (which was again validating everything already checked loads of times elsewhere!) is that it was opting to add a new billing account including new address and contact.

But as a duplicate it was also using the reason 'Strategic review of charges (SRoC)' and completing details for a new SROC charge information. So, when migrating we have opted to try and make the scenario more realistic. The new test selects the reason 'Licence transferred and now chargeable' and we stick with it being a new PRESROC charge information. We have plenty of coverage of the SROC process but none for the PRESROC. So, not only do we now cover creating a new billing contact, FAO and using the address lookup within the charge information flow, we also cover the PRESROC flow.

We also take it further than the previous test, approving the new charge information and checking the licence gets flagged for the next old charge scheme supplementary bill run.