Migrate legacy Bills tab test (#45)
We're onto the last folder of legacy tests; `internal/billing`. No surprise there are tests that have little to do with billing and would be better located elsewhere. The first of these was [mark-licence-for-supplementary-billing.spec.js](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/43).  This migrates `bills-tab.spec.js`.

We assumed this was checking you can see the bills tab and it has stuff much like `view-contacts.cy.js` which was checking you could see a licence's contacts. So, we moved `search.cy.js` and `view-contacts.cy.js` into a `licences/` folder and got ready to add this legacy test to them.

But when we started the migration we quickly realised it was just a simple test of permissions; whether a certain user should be able to see the bills tab.

We chose to leave the licence tests where they are. It made sense once we'd done it. But we've split the existing `bill-runs.spec.js` test into 2 and added them to `internal/user/permissions/` so it is much clearer what it is they are testing.