Migrate legacy billing tests plus some fixes (#46)
We're onto the last folder of legacy tests; `internal/billing`. No surprise there are tests that have little to do with billing and would be better located elsewhere. We've dealt with those so now we're on to the actual billing tests.

This change is all of them bar `internal/billing/billing-non-chargeable-licence-credits-back-historic-charges.spec.js` (we'll come back for that one!).

The first part of the migration was to get them working again. Some had started failing because of changes we've had to make to the billing journey to support SROC supplementary billing. Others it was because the test data being used is creating records that are too old and no longer lead to the expected behaviour. By correcting the tests and switching the test data sets used we've managed to get them up and running again.

The other issue was the complexity caused by many-layered describe blocks being used to note what is happening, when in real terms test frameworks see these as distinct tests. But also a switch to reusable functions that repeat the same actions. Only they make things more complex, brittle and in our case it was where all the breaks had happened.

So, we've done our usual. We've split things up to make what is being tested clearer. We've also simplified the journeys by removing the use of reusable functions.

We did add some new custom commands to support this that should prove useful in the future. We also corrected some tests (those in `internal/charge-information`) which had started failing. We'll need to keep an eye on them!