Update to support wal-dev-environment
**wal-dev-environment** is our repo which builds and maintains our local WRLS environment. This repo is now ready to be added to it but we need to make some tweaks. We're big fans of VSCode tasks and we have tasks in the **wal-dev-environment** we want to add that will trigger actions in this repo.

To do that though we need to tweak the `package.json` so that all environments are consistent; no more special case for `local`. To help with that though we also add in a drop down list. Now the tasks in this repo (should you wish to use it directly outside **wal-dev-environment**) are more intelligent and allow you to select your chosen environment from a drop down.