Migrate legacy non-chargeable licence credits (#49)
This is the last of the tests (!!) we found in `integration/internal/billing` that needed to be migrated. It is a test that checks if a licence is made non-chargeable (which is done by adding special charge information) then a credit is generated in the next supplementary bill run.

It requires us to first create, confirm and send a supplementary bill run, then add the charge information, and then create another supplementary bill run.

As such we struggled to determine where exactly it fitted within our current folder structure. We opted for a new one; `internal/billing/scenarios`. We know we will be adding more SROC billing scenarios in the very near future so it made sense to give them and this migrated test a home.

This change came in after support was added for multiple years to the SROC billing engine. We know it's broken some of our tests but we wanted to make sure this one was passing before merging. So, to support that and fixing the others we needed to be dynamic with what date to use in some of our screens. This change also adds a new custom command `currentFinancialYearDate()` which allows us to generate dates in the current financial year for use in our tests.