Add numberOfSrocBillingPeriods custom command
We recently merged in multi-year which broke tests because they assumed there would only be 1 bill in an SROC supplementary bill run.

We need the test suite to be more intelligent and work out how many bills there will be dynamically. This new command, when given the financial end year used for a bill run, can calculate how many bills we should expect to see.

We can then update our broken tests to use this.