Fix broken billing tests (#50)
https://eaflood.atlassian.net/browse/WATER-3984

We have updated our SROC supplementary billing engine to handle multiple years. When new charge information is added it will flag the licence for supplementary billing. Our first iteration of the engine only calculated bills for the financial year 2022-2023. But as a service, we support customers making changes as far back as 5 years ago. So, the supplementary billing engine needs to look at a licence's charge information and calculate what the charge would be, starting with the current financial year and working back 5 more.

> The SROC scheme into effect in April 2022, so the 5 years is cut off at that point. The PRESROC engine handles anything prior to this

What this results in is SROC supplementary bill runs will now feature multiple bills for the same invoice account, one for each year. Our tests were written when there would be only one. We need to amend our assertions, not only to account for this but to support the fact this number will keep increasing by one each year until we are 5 years away from 2022.

Fortunately, only one test is affected; `cypress/e2e/internal/billing/supplementary/journey.cy.js`. To fix it we focus on implementing a custom command that can calculate how many billing periods will be covered in a bill run when given a financial year. We then updated the test to focus on checking the number of bill runs rather than a £ value, as this will be more brittle and outside our control.

When implementing the changes we came up with a 'calculate once' pattern for the test. We move all the current calls to custom commands into the `beforeEach()` and then reference the values when needed. This works so well that we apply it to all tests using relevant custom commands.