Fix flaky tests where description keeps changing
We have been encountering an issue with our charge information tests. We have assertions that check for a specific description and purpose thinking that as we control the test data the descriptions should not change. But they do. Everything else in the tests is fine. It just appears that when the test data engine is creating the records the `charge_element` is using random purposes which means we cannot guarantee what we are asserting against.

Until we also have rebuilt that test data engine, we can't afford to include these assertions which make the tests flaky. Nothing is lost with their removal, the test of the journey is still valid.