Fix more flaky tests
Those with experience of acceptance tests will know they are extremely prone to [works on my machine](https://blog.codinghorror.com/the-works-on-my-machine-certification-program/).

We've started running these on different members machines and found they either because of performance or existing data setup that some tests will fail. Assertions on certain text content and whether the spinner page appears long enough to tests against are the primary contenders.

We also recently made a change to [Warn the user if a bill run is 'processing'](https://github.com/DEFRA/water-abstraction-ui/pull/2369). This made a structural change to the bill runs page that broke some tests.

These changes deal with those we've found so far. Excluding the new warning, when we've looked at the issue we've found we're asserting against something that doesn't really have much value. So, rather than make the tests more complex to cater for all possibilities we can just remove the assertion.