Fix tests not expecting date to be present
This was more a data issue. If a user had already been using the system, perhaps running tests manually, and these generates data for the current date we found tests would fail because they were not expecting to see the current date.

Having looked at the tests we have changed, we are happy there is enough left that we are still checking to confirm the bill run is no more.