Handle no bill runs (#56)
Another issue we've encountered getting the tests to run on everyone's machine is those folks who like to click the 'Delete all bills and charge information` button 😮😁

When they then run tests all the ones that test cancelling a bill run fail. This is because they have assertions to confirm the bill run doesn't exist in the results table. The problem is, if you've deleted all bill runs there is no results table!

Rather than force folks to have at least one bill run always present this change updates the tests affected to be able to handle both scenarios.

We use [conditional testing](https://docs.cypress.io/guides/core-concepts/conditional-testing) to achieve this. We are aware this is not considered a best practice as it means the test is no longer deterministic. But until we have a better test data seeding solution and are comfortable starting with a clean slate of a DB it's a compromise we have to accept.