Fix `cancel-existing` test (#62)
The `cancel-existing` test has started failing. Upon investigation, I've found that this is because the final check this test does after cancelling the bill runs is to check if the first bill run displayed on the Bill runs page is NOT a Supplementary bill run. This is not a good test as the bill run it checks is not created by the acceptance tests, so it could be anything.

Therefore, we will remove this assertion `.and('not.contain.text', 'Supplementary')` which is breaking the test. The assertion on the line above `.should('not.contain.text', 'Test Region')` is sufficient to check if the test bill runs have been cancelled.