Add SROC supplementary reissuing tests (#67)
https://eaflood.atlassian.net/browse/WATER-4061

Add a new reissue acceptance test to that covers selecting a bill for reissue, running the supplementary bill run, and confirming the bill is reissued as expected. We model this mainly on the existing presroc test, modifying it as necessary.

We also rename the existing presroc test to easily differentiate the two tests.