Fix improbable assertions on billing account no.
https://eaflood.atlassian.net/browse/WATER-4138
https://eaflood.atlassian.net/browse/WATER-4077

When running the tests recently 4 have been consistently failing. All featured assertions against a new billing account number that would be generated as part of the tests. Our confusion is how these ever worked!

The billing account number generated is based on existing data. But this can be different across the environments; what we do locally and in each of the non-prod environments mean we have no way of knowing what a new billing account number will be.

So, we remove the existing assertions which are causing the tests to fail.