Make send invitation more resilient and speedy
https://eaflood.atlassian.net/browse/WATER-4182
https://eaflood.atlassian.net/browse/WATER-4180

Whilst trying to sign off a release our acceptance tests kept failing. We finally tracked it down to the `cypress/e2e/internal/returns/send-invite.cy.js` test.

When we checked the test we found it isn't sending a return invite for a licence. Instead, it leaves the field where you enter a licence reference blank. This causes the service to generate an invite for _every_ licence with returns.

Obviously, this causes the test to be pretty slow. We don't need to do this to confirm the functionality is working plus it means the test can be broken easily. Licences are automatically updated daily by our import process and we often find 'duff' data creeps in.

That was the case this time. A licence with an invalid reference was imported which caused the send invite process to blow up every time it tried to schedule a notification for it.

If we enter the references for the licences we seed as part of testing, we'll still be testing the function works whilst speeding it up and removing duff data inadvertently breaking it.