Update bill run tests with journey changes (#78)
https://eaflood.atlassian.net/browse/WATER-4131

In WATER-4070 we logged that the Billing & Data team were struggling because they could not access the larger bill runs. The service would appear to be busy for a while and then eventually error.

We were able to figure out it was simply that the service was not able to return all the data needed using the current design, which listed out all the transactions in a bill run. We also knew the data could be collected in a more performant way.

So, we ended up rebuilding the whole journey of viewing a bill run, the bills in it and the transaction details.

Because the journey has changed, this has broken those tests that rely on the old journey. This change updates the tests to use the new journey.