Remove spinner checks from replace-cv-2023 test (#81)
https://eaflood.atlassian.net/browse/WATER-4336

Though running fine for one of the team our QA reported that this test appeared to be 'flaky'. When investigated it seems on the QA's machine there are times that the spinner page doesn't get a chance to appear before the bill run is 'sent'. This causes the test to fail, even though we are fine with that result!

This change removes the assertions for the spinner page in the send bill run process for `internal/billing/supplementary/replace-cv-2023-fin-year-no-changes.cy.js`.