Update bill test after decimal point being moved
https://eaflood.atlassian.net/browse/WATER-4358

We have just dealt with an issue reported by the billing and data team that the values shown against a transaction for the charge reference and supported source have the decimal in the wrong place.

We've fixed the code but as expected this has broken a test. This change updates it to match what the UI is now displaying.