Update create bill run journeys
https://eaflood.atlassian.net/browse/WATER-4375

In order to implement changes to support new two-part tariff bill run options we needed to update the create bill run journey. We took advantage of this to migrate the functionality into [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system).

However, this means the elements we are referencing in these tests no longer exist.

This updates the billing tests to use the new element references.