Remove reissue tests
We no longer trigger the reissue process as part of supplementary billing. We are now 7 months on from SROC supplementary billing going live and reissue still hasn't been enabled in production.

The current position is it will never be but we are just waiting on Billing & Data making an official decision on this.

In the meantime we have disregarded reissue when migrating existing features due to the complexity it adds. We also needed to make changes to supplementary to support [Exclude Financial Year from Supplementary Billing where Annual Bill has not been run for region & year](https://eaflood.atlassian.net/browse/WATER-4403). To support this we've siloed reissue again because of the complexity it is adding to the process.

In the unlikely event that reissue is still required we can bring back these tests. But based on current version of the service this will no longer pass.