Update billing scenarios due to changes (#85)
https://eaflood.atlassian.net/browse/WATER-4375

To implement changes to support new two-part tariff bill run options we needed to update the create bill run journey. We took advantage of this to migrate the functionality into [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system).

We also migrated the cancel bill run, send bill run and remove a bill and licence screens.

However, this means the elements we are referencing in these tests no longer exist.

This updates the billing tests to use the new element references.

---

One of the key changes in the migrated journey is we now return users to the bill runs page rather than redirect them to a spinner page. Hence a number of the changes are about waiting for a bill run to change from `BUILDING` to a finished state.

We've also removed the cancel button on the spinner page (should you get there). We've found this process is flawed in the legacy service and introduces issues. So, cancelling an in-progress bill run is no longer possible hence the tests have gone.

The other key change is the removal of the reissue tests. We no longer trigger the reissue process as part of supplementary billing. We are now 7 months on from SROC supplementary billing going live and reissue still hasn't been enabled in `production`.

The current position is it will never be but we are just waiting on official confirmation.

In the meantime we have disregarded reissue when migrating existing features due to the complexity it adds. We also needed to make changes to supplementary to support [Exclude Financial Year from Supplementary Billing where Annual Bill has not been run for region & year](https://eaflood.atlassian.net/browse/WATER-4403). To aid this, we've siloed reissue again because of the complexity it is adding to the process.

In the unlikely event that reissue is still required, we can bring back these tests. But based on the current version of the service these tests will no longer pass.