Supp. journey fixed
The issue is that having moved into 2024-25 we've hit a point where part of the PRESROC charge version is no longer included in the bill run.

It starts on Jan 1 2019. When the billing period was 2023-24 five years prior to that year would include the billing period 2018-19. This means there would be 4 supplementary bills with the first just being for the period 2019-01-01 to 2019-03-31.

Now the billing period is 2024-25 the period 2018-2019 has dropped off. This means 1 less bill which also effects the bill run total.