Fix broken annual billing
We had overlooked that the annual billing tests were using the `sroc-billing-data` fixture as well. Adding an annual bill run to that fixture causes the annual test to fail due to there be an existing bill run.

It was a simple fix in [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service). We just needed to add in a copy of the fixture set without the annual bill run. But to make things clearer we've renamed the 3 sets hence the need to go back through all our tests and update the references.