Add test for supplementary bump
This new test gives us a means to check that the supplementary end year bump feature is working. It uses a new SROC fixture that is the same as our normal one only the annual bill run created is for the previous billing period.

This means we can confirm when we create an SROC supplementary that its financial year end is also for the previous billing period.